Haiku: add some locking.
Unlike most systems, Haiku runs a separate thread for each window. This
can lead to problems, as part of the IUP code is run in the main
application thread, and other parts are run in window specific threads.

Lock the window thread whenever accessing UI elements, so that it's
possible to get things working fromthe main thread. This may not be
enough for real applications, but it's ok to get at least the tests
running.